stats: change the type of 'size' parameter to 'size_t'
This is to disallow negative values. The size of the array should be
a positive number only.

+++ c/apps/stats.h
@@ -1,6 +1,8 @@
 #ifndef STATS_H_
 #define STATS_H_
 
+#include <stddef.h>
+
 /**
  * Calculates the minimum value in the specified array, or NAN if the
  * array is empty.
@@ -9,7 +11,7 @@
  * @param size The array size.
  * @param res  The output where the result is stored.
  */
-int stats_min(double const *data, int size, double *res);
+int stats_min(double const *data, size_t size, double *res);
 
 /**
  * Calculates the maximum value in the specified array, or NAN if the
@@ -19,7 +21,7 @@ int stats_min(double const *data, int size, double *res);
  * @param size The array size.
  * @param res  The output where the result is stored.
  */
-int stats_max(double const *data, int size, double *res);
+int stats_max(double const *data, size_t size, double *res);
 
 /**
  * Calculates the sum of all values in the specified array.
@@ -28,7 +30,7 @@ int stats_max(double const *data, int size, double *res);
  * @param size The array size.
  * @param res  The output where the result is stored.
  */
-int stats_sum(double const *data, int size, double *res);
+int stats_sum(double const *data, size_t size, double *res);
 
 /**
  * Calculates the arithmetic mean of the values in the specified array.
@@ -37,7 +39,7 @@ int stats_sum(double const *data, int size, double *res);
  * @param size The array size.
  * @param res  The output where the result is stored.
  */
-int stats_avg(double const *data, int size, double *res);
+int stats_avg(double const *data, size_t size, double *res);
 
 /**
  * Calculates the variance based on a sample of the values in the
@@ -47,7 +49,7 @@ int stats_avg(double const *data, int size, double *res);
  * @param size The array size.
  * @param res  The output where the result is stored.
  */
-int stats_var(double const *data, int size, double *res);
+int stats_var(double const *data, size_t size, double *res);
 
 /**
  * Calculates the standard deviation based on a sample of the values in the
@@ -57,6 +59,6 @@ int stats_var(double const *data, int size, double *res);
  * @param size The array size.
  * @param res  The output where the result is stored.
  */
-int stats_stdev(double const *data, int size, double *res);
+int stats_stdev(double const *data, size_t size, double *res);
 
 #endif /* STATS_H_ */
